requirements.txt
absl-py==1.4.0
aiohttp==3.9.3
aioice==0.9.0
aiortc==1.7.0
aiosignal==1.3.1
altair==5.2.0
amqp==5.2.0
annotated-types==0.6.0
ansi2html==1.8.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
av==11.0.0
Babel==2.14.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
billiard==4.2.0
bleach==6.1.0
blinker==1.6.2
blis==0.7.11
build==1.0.3
cachetools==5.3.1
catalogue==2.0.10
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chart-studio==1.1.0
chroma-hnswlib==0.7.3
chromadb==0.4.22
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudpathlib==0.16.0
cmake==3.26.4
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.1
comtypes==1.2.0
confection==0.1.4
contourpy==1.0.7
cryptography==42.0.5
cycler==0.11.0
cymem==2.0.8
dash==2.13.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dataclasses-json==0.6.4
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
distlib==0.3.7
distro==1.9.0
Django==5.0.4
django-celery-results==2.5.1
djangorestframework==3.14.0
dnspython==2.6.1
EasyProcess==1.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
entrypoint2==1.1
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==2.0.1
fanstatic==1.3
fastapi==0.109.2
fastjsonschema==2.19.1
filelock==3.12.2
Flask==2.2.5
Flask-SQLAlchemy==3.1.1
flatbuffers==23.5.26
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2023.12.2
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.41
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-crc32c==1.5.0
google-pasta==0.2.0
googleapis-common-protos==1.62.0
GPUtil==1.4.0
greenlet==3.0.3
grpcio==1.58.0
h11==0.14.0
h5py==3.9.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.20.3
humanfriendly==10.0
idna==3.6
ifaddr==0.2.0
imageio==2.33.1
importlib-metadata==6.11.0
importlib-resources==6.1.1
ipykernel==6.29.3
ipython==8.22.1
ipython-genutils==0.2.0
ipywidgets==8.0.2
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.2.0
js==1.0
json5==0.9.17
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.9.0
jupyter-lsp==2.2.3
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.1.2
jupyterlab-widgets==3.0.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.3
keras==2.15.0
keyboard==0.13.5
kiwisolver==1.4.4
kombu==5.3.5
kubernetes==29.0.0
langchain==0.1.9
langchain-community==0.0.24
langchain-core==0.1.27
langchain-openai==0.0.7
langcodes==3.3.0
langdetect==1.0.9
langsmith==0.1.9
lazy_loader==0.3
libclang==16.0.6
llvmlite==0.41.1
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.0
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.2.0
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
mss==9.0.1
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
mysql-connector-python==8.0.29
nbclassic==0.4.8
nbclient==0.9.0
nbconvert==7.16.1
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
notebook==7.1.1
notebook_shim==0.2.4
numba==0.58.1
numpy==1.24.3
oauthlib==3.2.2
onnxruntime==1.16.3
openai==1.12.0
opencv-python==4.9.0.80
openpyxl==3.1.2
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
opt-einsum==3.3.0
orjson==3.9.15
outcome==1.2.0
overrides==7.7.0
packaging==23.2
pandas==1.5.3
pandocfilters==1.5.1
parso==0.8.3
patsy==0.5.3
pickleshare==0.7.5
pillow==10.2.0
platformdirs==3.11.0
plotly==5.16.1
pooch==1.8.0
posthog==3.4.1
preshed==3.0.9
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.24.3
psutil==5.9.8
psycopg2==2.9.9
pulsar-client==3.4.0
pure-eval==0.2.2
pyarrow==14.0.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.13
pycparser==2.21
pydantic==2.6.2
pydantic_core==2.16.3
pydeck==0.8.1b0
pydot==1.4.2
pyee==11.1.0
Pygments==2.17.2
pylibsrtp==0.10.0
PyMatting==1.1.12
pyOpenSSL==24.0.0
pyparsing==3.0.9
PyPika==0.48.9
pypiwin32==223
pyproject_hooks==1.0.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
pyreadline3==3.4.1
pyrsistent==0.19.2
pyscreenshot==3.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-env==1.0.0
python-json-logger==2.0.7
pyttsx3==2.90
pytz==2022.7.1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.4.0
QtPy==2.3.0
referencing==0.33.0
regex==2023.12.25
rembg==2.0.53
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.18.0
rsa==4.9
scikit-image==0.22.0
scikit-learn==1.2.2
scikit-plot==0.3.7
scipy==1.10.1
seaborn==0.12.2
selenium==4.12.0
Send2Trash==1.8.2
six==1.16.0
sklearn==0.0.post5
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.10.0
SQLAlchemy==2.0.27
sqlparse==0.4.4
srsly==2.4.8
stack-data==0.6.3
starlette==0.36.3
statsmodels==0.14.0
streamlit==1.33.0
streamlit-lottie==0.0.5
streamlit-webrtc==0.47.6
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.15.2
tensorboard-data-server==0.7.1
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.18.0
textblob==0.17.1
thinc==8.2.3
threadpoolctl==3.1.0
tifffile==2023.12.9
tiktoken==0.6.0
tinycss2==1.2.1
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.1.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
trio==0.22.2
trio-websocket==0.10.3
typer==0.9.0
types-python-dateutil==2.8.19.20240106
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
uri-template==1.3.0
urllib3==1.26.18
uvicorn==0.27.1
validators==0.22.0
vine==5.1.0
virtualenv==20.24.3
wasabi==1.1.2
watchdog==3.0.0
watchfiles==0.21.0
wcwidth==0.2.13
weasel==0.3.4
webcolors==1.13
webdriver-manager==4.0.0
webencodings==0.5.1
WebOb==1.8.7
websocket-client==1.7.0
websockets==12.0
Werkzeug==2.2.3
widgetsnbextension==4.0.3
wikipedia==1.4.0
wordcloud==1.9.3
wordninja==2.0.0
wrapt==1.14.1
wsproto==1.2.0
xgboost==2.0.3
xlrd==2.0.1
yarl==1.9.4
zipp==3.17.0
